refactor(schema): address review comments on PR #438
- Rename monochange_schema_gen to xtask, use clap CLI, add cargo xtask alias

- Make Raw* types pub(crate), add schema modules in monochange_config and monochange_core

- Use #[schemars(rename)] on Raw* types for  naming

- Move inline tests to __tests__/lib_tests.rs

- Update devenv.nix to use cargo xtask schema update/check

- Simplify post_process_config by removing manual  renaming
