+++ .cargo/config.toml
@@ -4,3 +4,4 @@ insta = ["bin", "cargo-insta"]
 llvm-cov = ["bin", "cargo-llvm-cov"]
 nextest = ["bin", "cargo-nextest"]
 semver-checks = ["bin", "cargo-semver-checks"]
+xtask = ["run", "--package", "xtask", "--"]
