+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["crates/*", "docs", "crates/monochange_schema_gen"]
+members = ["crates/*", "docs"]
 resolver = "2"
 
 [workspace.package]
