+++ docs/src/schemas/release-record.v0.1.schema.json
@@ -850,7 +850,7 @@
 			"type": "array"
 		},
 		"schemaVersion": {
-			"const": "0.1",
+			"default": "0.1",
 			"type": "string"
 		},
 		"updatedChangelogs": {
